--- CONCEPTS.md
@@ -1,3 +1,0 @@
-# Concepts
-
-## Source
