+++ core/core/infra/stream_processor.py
