+++ core/tests/scenarios/test_phishing.py
@@ -1,11 +1,13 @@
+import functools
 import collections
 import pytest
 from fastapi.testclient import TestClient
 from sqlmodel import Session, SQLModel, create_engine
 from sqlmodel.pool import StaticPool
 
-from api.main import app
-from api.db import engine, get_session
+from core import infra
+from core.main import app
+from core.db import engine, get_session
 
 
 @pytest.fixture(name="session")
@@ -32,12 +34,14 @@ def client_fixture(session: Session):
 
 def test_phishing(client: TestClient):
 
+    message_bus = infra.DummyMessageBus()
+
     # Create source
     response = client.post(
-        "/api/sources",
+        "/api/message_bus",
         json={"name": "foo", "protocol": "DUMMY"},
     )
     assert response.status_code == 200
-    assert len(client.get("/api/sources/").json()) == 1
-    assert client.get("/api/sources/1").json()["name"] == "foo"
-    assert client.get("/api/sources/1").json()["protocol"] == "DUMMY"
+    assert len(client.get("/api/message_bus/").json()) == 1
+    assert client.get("/api/message_bus/1").json()["name"] == "foo"
+    assert client.get("/api/message_bus/1").json()["protocol"] == "DUMMY"
